requirements.txt
# 基本爬蟲相關套件
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# 數據處理相關
pandas==1.5.3
numpy==1.24.3

# 數據庫連接
mysql-connector-python==8.2.0
SQLAlchemy>=1.4.24,<2.0.0  # 修改為與 Airflow 相容的版本
pymysql==1.1.0

# Airflow 相關
apache-airflow[google,mysql]==2.7.1

# GCP 相關
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.13.0
google-auth==2.23.4

# 其他工具
python-dotenv==1.0.0
urllib3<2.0.0  # 降級 urllib3 以避免可能的相容性問題
